+++ src/popups/StopPlacePopup/StopPlacePopup.js
@@ -1,4 +1,10 @@
-import React, { useCallback, useEffect, useMemo, useState } from "react";
+import React, {
+  Fragment,
+  useCallback,
+  useEffect,
+  useMemo,
+  useState,
+} from "react";
 import PropTypes from "prop-types";
 import Feature from "ol/Feature";
 import { useSelector } from "react-redux";
@@ -62,37 +68,40 @@ function StopPlacePopup({ feature }) {
 
   const renderContent = useCallback(
     ([key, value]) => {
-      if (key === "ticketMachine") return null;
-      if (key === "alternativeTransport") {
-        return (
-          !/^NO|UNKNOWN$/.test(value?.state) && (
+      let content = value && (
+        <div key={key}>
+          <div>
+            {t(key)}: {t(formatYesNoData(value))}
+          </div>{" "}
+          <br />
+        </div>
+      );
+
+      if (value && typeof value === "object") {
+        const entries = Object.entries(value);
+        if (entries.length) {
+          content = (
             <div key={key}>
               <fieldset>
                 <legend>{t(key)}</legend>
                 <br />
-                <div>{t("Shuttle-Fahrdienst")}</div>
-                {value.note?.[i18n.language] && (
-                  <>
-                    <br />
-                    <div>{value.note[i18n.language]}</div>
-                  </>
-                )}
-                <br />
+                <div>{entries.map(renderContent)}</div>
               </fieldset>
               <br />
             </div>
-          )
-        );
+          );
+        }
       }
 
-      if (key === "accessibility") {
-        return (
+      if (key === "ticketMachine") content = null;
+      if (key === "alternativeTransport") {
+        content = !/^NO|UNKNOWN$/.test(value?.state) && (
           <div key={key}>
             <fieldset>
               <legend>{t(key)}</legend>
               <br />
-              <div>{t(formatYesNoData(value?.state))}</div>
-              {value?.state === "NO" && value.note?.[i18n.language] && (
+              <div>{t("Shuttle-Fahrdienst")}</div>
+              {value.note?.[i18n.language] && (
                 <>
                   <br />
                   <div>{value.note[i18n.language]}</div>
@@ -105,26 +114,43 @@ function StopPlacePopup({ feature }) {
         );
       }
 
-      if (value && key === "passengerInformation") {
-        const entries = Object.entries(value);
-        return (
+      if (key === "accessibility") {
+        content = (
           <div key={key}>
             <fieldset>
               <legend>{t(key)}</legend>
               <br />
-              {entries.map(([k, val]) => {
+              <div>{t(formatYesNoData(value?.state))}</div>
+              {value?.note && (
+                <>
+                  <br />
+                  {typeof value.note === "object"
+                    ? value.note[i18n.language] || value.note.de
+                    : value.note}
+                </>
+              )}
+              <br />
+            </fieldset>
+            <br />
+          </div>
+        );
+      }
+
+      if (value && key === "passengerInformation") {
+        const entries = Object.entries(value).filter(([k, val]) => {
+          return val !== "NO" && val !== "UNKNOWN" && k !== "note";
+        });
+        content = !!entries.length && (
+          <div key={key}>
+            <fieldset>
+              <legend>{t(key)}</legend>
+              <br />
+              {entries.map(([k]) => {
                 return (
-                  val !== "NO" &&
-                  val !== "UNKNOWN" &&
-                  k !== "note" && (
-                    <>
-                      <div>
-                        {t(k)}
-                        {val === "UNKNOWN" && `: ${val}`}
-                      </div>
-                      <br />
-                    </>
-                  )
+                  <Fragment key={k}>
+                    <div>{t(k)}</div>
+                    <br />
+                  </Fragment>
                 );
               })}
             </fieldset>
@@ -134,9 +160,11 @@ function StopPlacePopup({ feature }) {
       }
 
       if (value && key === "note") {
-        return (
+        content = (
           <div key={key}>
-            {typeof value === "object" ? value[i18n.language] : value}
+            {typeof value === "object"
+              ? value[i18n.language] || value.de
+              : value}
             <br />
           </div>
         );
@@ -152,7 +180,7 @@ function StopPlacePopup({ feature }) {
           console.log("StopPlacePopup. Not a parseable url: ", value);
           niceVal = value;
         }
-        return (
+        content = (
           <div key={key}>
             <Link href={url}>{niceVal}</Link>
             <br />
@@ -160,47 +188,23 @@ function StopPlacePopup({ feature }) {
         );
       }
 
-      if (value && typeof value === "object") {
-        const entries = Object.entries(value);
-        if (entries.length) {
-          return (
-            <div key={key}>
-              <fieldset>
-                <legend>{t(key)}</legend>
-                <br />
-                <div>{entries.map(renderContent)}</div>
-              </fieldset>
-              <br />
-            </div>
-          );
-        }
-      }
-
-      return (
-        value && (
-          <div key={key}>
-            <div>
-              {t(key)}: {t(formatYesNoData(value))}
-            </div>{" "}
-            <br />
-          </div>
-        )
-      );
+      return content || null;
     },
     [i18n.language, t],
   );
+
+  const popupContent = useMemo(() => {
+    const content = Object.entries(data?.prmInformation || {})
+      .map(renderContent)
+      .filter(Boolean);
+    return content?.length ? content : t(`Keine Daten für diese Station`);
+  }, [data?.prmInformation, renderContent, t]);
 
   if (loading) {
     return <div>Loading ...</div>;
   }
 
-  return (
-    <div>
-      {data?.prmInformation
-        ? Object.entries(data?.prmInformation || {}).map(renderContent)
-        : `No data for this station`}
-    </div>
-  );
+  return <div>{popupContent}</div>;
 }
 
 StopPlacePopup.propTypes = propTypes;