+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.25.2](https://github.com/geops/trafimage-maps/compare/v1.25.1...v1.25.2) (2024-01-08)
+
+### Bug Fixes
+
+- **Regionenkarte:** move highlight logic from popup to RegionenkarteLayer, improve Select UI ([#1126](https://github.com/geops/trafimage-maps/issues/1126)) ([74ee65c](https://github.com/geops/trafimage-maps/commit/74ee65c2b99edbd752c553552f10bb77e31f5e46))
+
+### [1.25.1](https://github.com/geops/trafimage-maps/compare/v1.25.0...v1.25.1) (2024-01-04)
+
+### Bug Fixes
+
+- **handicap:** improve AMO highlighting, remove datalinks and dataservice from topic info ([#1125](https://github.com/geops/trafimage-maps/issues/1125)) ([0e766c5](https://github.com/geops/trafimage-maps/commit/0e766c535d24a0363528ac63d80c7c5d200a8aa4))
+
+## [1.25.0](https://github.com/geops/trafimage-maps/compare/v1.24.2...v1.25.0) (2024-01-02)
+
+### Features
+
+- **Handicap:** use AMO data for handicap layers ([#1122](https://github.com/geops/trafimage-maps/issues/1122)) ([b7b6fe3](https://github.com/geops/trafimage-maps/commit/b7b6fe3c433e7b69b688b12816eefdecaa23f50e))
+
 ### [1.24.2](https://github.com/geops/trafimage-maps/compare/v1.24.1...v1.24.2) (2023-12-21)
 
 ### Bug Fixes